requirements.txt
Flask
wordsegment
pyspellchecker
flask-restx
importlib-metadata; python_version < "3.8"
